requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
youtube-transcript-api==1.2.2
python-dotenv==1.0.0
